permission file added to project
take values from participant info

only update metadata if object has isMirrored Prop on it

y# This is a combination of 3 commits.

send meta if property exists or send prevMetadata

added condition to manage permission role

added comments for better code readbility
