+++ src/core/webrtc/room-context.ts
@@ -19,7 +19,7 @@ import * as LiveKitServer from '@api.stream/livekit-server-sdk'
 import decode from 'jwt-decode'
 import { log, CoreContext } from '../context'
 import { Role } from '../types'
-import { hasPermission, Permission } from '../../logic'
+import { hasPermission, Permission } from '../../helpers/permission'
 export * as Livekit from 'livekit-client'
 
 /**
@@ -401,6 +401,8 @@ export class RoomContext implements LSRoomContext {
           case DataType.ChatMessage: {
             return this._appendChat(payload, participant, kind)
           }
+          
+          /* Updating the participant metadata. */
           case DataType.ParticipantMetadataUpdate: {
             const strData = decoder.decode(payload)
             const data: ParticipantDataObject = JSON.parse(strData)
@@ -471,19 +473,28 @@ export class RoomContext implements LSRoomContext {
       this.livekitRoom = null
       this._updateParticipants()
     })
+    
+    /* The above code is subscribing to the RoomEvent.ParticipantMetadataChanged event. When the event is
+    triggered, the code checks if the metadata has changed. If it has, it parses the metadata and checks
+    if the participant has the permission to manage themselves. If they do, it updates the
+    guestParticipantsStore and the participants. */
     this.subscribeToRoomEvent(
       RoomEvent.ParticipantMetadataChanged,
-      (metadata: string, participant: Participant, prevMetadata: string) => {
-        const meta = JSON.parse(metadata)
-        if (hasPermission(meta?.participantRole, Permission.ManageSelf)) {
-          const data = {
-            participantId: participant?.identity,
-            metadata: meta,
-            type: DataType.ParticipantMetadataUpdate,
-          } as ParticipantDataObject
-          this._updateGuestParticipantsStore(data)
-          this._updateParticipants()
-          return
+      (metadata: string, participant: Participant) => {
+        if (metadata !== participant?.metadata) {
+          const meta = JSON.parse(participant?.metadata)
+          if (
+            hasPermission(meta?.participantRole, Permission.ManageSelf) 
+          ) {
+            const data = {
+              participantId: participant?.identity,
+              metadata: meta.hasOwnProperty('isMirrored') ? meta : JSON.parse(metadata),
+              type: DataType.ParticipantMetadataUpdate,
+            } as ParticipantDataObject
+            this._updateGuestParticipantsStore(data)
+            this._updateParticipants()
+            return
+          }
         }
       },
     )
@@ -551,6 +562,8 @@ export class RoomContext implements LSRoomContext {
       const remotes = Array.from(this.livekitRoom.participants.values())
       const parts = [this.livekitRoom.localParticipant] as Participant[]
       parts.push(...remotes)
+
+      /* Updating the metadata of the participants in the room. */
       const updatedParts = parts.map((participant) => {
         const existingGuestParticipantMetadata =
           this.guestParticipantMetadata.find(
@@ -566,6 +579,7 @@ export class RoomContext implements LSRoomContext {
         }
         return participant
       })
+      
       this.participants = updatedParts
     }
   }
